种地

+++ Assets/Editor/Window/NpcCreateWindowEditor.cs
@@ -727,7 +727,7 @@ public class NpcCreateWindowEditor : EditorWindow
                 textInfoService.DeleteDataById(TextEnum.Talk, itemTalkInfo.id);
             }
             GUILayout.Label("talkId：", GUILayout.Width(100), GUILayout.Height(20));
-            itemTalkInfo.id = int.Parse(EditorGUILayout.TextArea(itemTalkInfo.id + "", GUILayout.Width(150), GUILayout.Height(20)));
+            itemTalkInfo.id = long.Parse(EditorGUILayout.TextArea(itemTalkInfo.id + "", GUILayout.Width(150), GUILayout.Height(20)));
             itemTalkInfo.text_id = itemTalkInfo.id;
             GUILayout.Label("对话顺序：", GUILayout.Width(100), GUILayout.Height(20));
             itemTalkInfo.text_order = int.Parse(EditorGUILayout.TextArea(itemTalkInfo.text_order + "", GUILayout.Width(50), GUILayout.Height(20)));
